split non/kernel chainparamsbase
Move the functions not using the ArgsManager to the kernel.

Also moves the globalChainBaseParams to kernel, in order to move the
BaseParams call to the kernel. This lets validation.cpp depend on on one
fewer gArgs aware component, by linking with the kernel chainparams.

Add a SetGlobalBaseParams function in order to allow non-kernel code to
set the value of the globalChainBaseParams.
